[Refactor] #4025 rooms-pit-nest.cpp/h をpit-nest-util.cpp/h にリネームした
