+++ Hengband/Hengband/Hengband.vcxproj
@@ -1719,7 +1719,7 @@
     <ClCompile Include="..\..\src\room\rooms-city.cpp" />
     <ClCompile Include="..\..\src\room\rooms-fractal.cpp" />
     <ClCompile Include="..\..\src\room\rooms-normal.cpp" />
-    <ClCompile Include="..\..\src\room\rooms-pit-nest.cpp" />
+    <ClCompile Include="..\..\src\room\pit-nest-util.cpp" />
     <ClCompile Include="..\..\src\room\rooms-special.cpp" />
     <ClCompile Include="..\..\src\room\rooms-trap.cpp" />
     <ClCompile Include="..\..\src\room\rooms-vault.cpp" />
@@ -1997,7 +1997,7 @@
     <ClInclude Include="..\..\src\room\rooms-city.h" />
     <ClInclude Include="..\..\src\room\rooms-fractal.h" />
     <ClInclude Include="..\..\src\room\rooms-normal.h" />
-    <ClInclude Include="..\..\src\room\rooms-pit-nest.h" />
+    <ClInclude Include="..\..\src\room\pit-nest-util.h" />
     <ClInclude Include="..\..\src\room\rooms-special.h" />
     <ClInclude Include="..\..\src\room\rooms-trap.h" />
     <ClInclude Include="..\..\src\room\rooms-vault.h" />
