+++ src/Makefile.am
@@ -780,6 +780,7 @@ hengband_SOURCES = \
 	room/cave-filler.cpp room/cave-filler.h \
 	room/door-definition.cpp room/door-definition.h \
 	room/lake-types.h room/room-types.h \
+	room/pit-nest-util.cpp room/pit-nest-util.h \
 	room/room-generator.cpp room/room-generator.h \
 	room/room-info-table.cpp room/room-info-table.h \
 	room/rooms-builder.cpp room/rooms-builder.h \
@@ -789,7 +790,6 @@ hengband_SOURCES = \
 	room/rooms-nest.cpp room/rooms-nest.h \
 	room/rooms-normal.cpp room/rooms-normal.h \
 	room/rooms-pit.cpp room/rooms-pit.h\
-	room/rooms-pit-nest.cpp room/rooms-pit-nest.h \
 	room/rooms-special.cpp room/rooms-special.h \
 	room/rooms-trap.cpp room/rooms-trap.h \
 	room/rooms-vault.cpp room/rooms-vault.h \
