+++ src/room/pit-nest-util.h
