+++ src/room/rooms-pit.cpp
@@ -12,8 +12,8 @@
 #include "monster/monster-list.h"
 #include "monster/monster-util.h"
 #include "room/door-definition.h"
+#include "room/pit-nest-util.h"
 #include "room/space-finder.h"
-#include "room/rooms-pit-nest.h"
 #include "system/dungeon-info.h"
 #include "system/floor-type-definition.h"
 #include "system/grid-type-definition.h"
